fix: develop branch build오류 해결
repository id부분 Long타입으로 변경 -> main, test 오류 해결

+++ backend/lighthouse/src/main/java/com/ssafy/lighthouse/domain/study/service/StudyMaterialServiceImpl.java
@@ -46,7 +46,7 @@ public class StudyMaterialServiceImpl implements StudyMaterialService {
 
 	@Override
 	public StudyMaterial findById(final Long id) {
-		final Optional<StudyMaterial> studyMaterial = studyMaterialRepository.findById(id.intValue());
+		final Optional<StudyMaterial> studyMaterial = studyMaterialRepository.findById(id);
 		studyMaterial.orElseThrow(() -> new StudyMaterialNotFoundException(id));
 		return studyMaterial.get();
 	}
